Вносим правки после ревью

+++ commerce/delivery/src/main/java/ru/yandex/practicum/controller/DeliveryController.java
@@ -8,6 +8,7 @@ import ru.yandex.practicum.model.DeliveryDto;
 import ru.yandex.practicum.model.OrderDto;
 import ru.yandex.practicum.service.DeliveryService;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Slf4j
@@ -41,7 +42,7 @@ public class DeliveryController implements DeliveryOperations {
     }
 
     @Override
-    public Double calculateDeliveryCost(OrderDto orderDto) {
+    public BigDecimal calculateDeliveryCost(OrderDto orderDto) {
         log.info("Расчет стоимости доставки: {}", orderDto);
         return deliveryService.calculateDeliveryCost(orderDto);
     }
