+++ commerce/interaction-api/dto/src/main/java/ru/yandex/practicum/model/PaymentDto.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class PaymentDto {
     private String paymentId;
-    private Double totalPrice;
-    private Double deliveryTotal;
-    private Double feeTotal;
+    private BigDecimal totalPrice;
+    private BigDecimal deliveryTotal;
+    private BigDecimal feeTotal;
 }
