+++ commerce/interaction-api/dto/src/main/java/ru/yandex/practicum/request/NewProductInWarehouseRequest.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.yandex.practicum.model.DimensionDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -23,6 +24,5 @@ public class NewProductInWarehouseRequest {
 
     @NotNull
     @DecimalMin(value = "1")
-    private Double weight;
-
+    private BigDecimal weight;
 }
