+++ commerce/interaction-api/feign/src/main/java/ru/yandex/practicum/PaymentOperations.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.model.OrderDto;
 import ru.yandex.practicum.model.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentOperations {
@@ -13,13 +14,13 @@ public interface PaymentOperations {
     PaymentDto payment(@Valid @RequestBody OrderDto order);
 
     @PostMapping("/api/v1/payment/totalCost")
-    Double totalCost(@Valid @RequestBody OrderDto order);
+    BigDecimal totalCost(@Valid @RequestBody OrderDto order);
 
     @PostMapping("/api/v1/payment/refund")
     void refund(@RequestBody UUID paymentId);
 
     @PostMapping("/api/v1/payment/productCost")
-    Double productCost(@Valid @RequestBody OrderDto order);
+    BigDecimal productCost(@Valid @RequestBody OrderDto order);
 
     @PostMapping("/api/v1/payment/failed")
     void paymentFailed(@RequestBody UUID paymentId);
