+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,16 +3,17 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.model.OrderDto;
 import ru.yandex.practicum.model.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
     PaymentDto payment(OrderDto order);
 
-    Double totalCost(OrderDto order);
+    BigDecimal totalCost(OrderDto order);
 
     void refund(UUID paymentId);
 
-    Double productCost(OrderDto order);
+    BigDecimal productCost(OrderDto order);
 
     void paymentFailed(UUID paymentId);
 }
